moved apolloclient attaching logic up
