+++ src/app/App.tsx
@@ -1,11 +1,11 @@
 import { useAuth } from "@saleor/sdk";
 import { useRouter } from "next/router";
-import React, { useEffect } from "react";
+import React from "react";
 
 import { Loader } from "@components/atoms";
 import { useDynamicRouteRedirect } from "@hooks";
 import { demoMode } from "@temp/constants";
-import { getSaleorApi, ShopConfig } from "@utils/ssr";
+import { ShopConfig } from "@utils/ssr";
 
 import {
   Footer,
@@ -21,12 +21,6 @@ import "../globalStyles/scss/index.scss";
 
 type AppProps = ShopConfig;
 
-declare global {
-  interface Window {
-    __APOLLO_CLIENT__: any;
-  }
-}
-
 const App: React.FC<AppProps> = ({
   footer,
   mainMenu,
@@ -37,15 +31,6 @@ const App: React.FC<AppProps> = ({
   const willRedirect = useDynamicRouteRedirect();
   const { tokenRefreshing, tokenVerifying } = useAuth();
   const loading = tokenRefreshing || tokenVerifying || willRedirect;
-
-  useEffect(() => {
-    const attachClient = async () => {
-      const { apolloClient } = await getSaleorApi();
-      if (window) window.__APOLLO_CLIENT__ = apolloClient;
-    };
-    if (process.env.NEXT_PUBLIC_ENABLE_APOLLO_DEVTOOLS === "true")
-      attachClient();
-  }, []);
 
   return (
     <ShopProvider shopConfig={shopConfig}>
